Added two new functions to the channel api and reorganized tests accordingly.
